network_pipeline/pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "network-pipeline"
version = "0.1.0"
description = "Incremental Meshed network pipeline slices."
requires-python = ">=3.10"
dependencies = ["PyYAML>=6,<7"]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
